+++ src/qt/Ask_For_Pass/ask_for_pass_qt.cpp
@@ -119,7 +119,7 @@ extern "C" unsigned int maxDataSize(ppk_data_type type)
 	switch(type)
 	{
 		case ppk_string:
-			return -1;
+			return (unsigned int)-1;
 		case ppk_blob:
 			return 0;
 	}
