+++ src/windows/Ask_For_Pass/ask_for_pass_win32.cpp
@@ -123,7 +123,7 @@ extern "C" unsigned int maxDataSize(ppk_data_type type)
 	switch(type)
 	{
 		case ppk_string:
-			return -1;
+			return (unsigned int)-1;
 		case ppk_blob:
 			return 0;
 	}
